added titles fixed vaults not showing on home

+++ Keepr.client/src/components/KeepModal.vue
@@ -3,28 +3,29 @@
     <template #modal-body>
       <div class="container">
         <div class="row">
-          <div class="col-md-6 order-2">
+          <div class="col-md-6 order-1">
             <img
               class="img-fluid keep-img rounded"
               :src="keep.img"
               alt="https://thiscatdoesnotexist.com"
             />
           </div>
-          <div class="col-md-6 order-1 order-md-2">
+          <div class="col-md-6 order-2 order-md-1">
             <div class="ms-4 d-flex flex-column justify-content-center">
               <div class="text-end">
                 <button
+                  title="Close Modal"
                   type="button"
                   class="btn-close p-2"
                   data-bs-dismiss="modal"
                   aria-label="Close"
                 ></button>
               </div>
-              <div class="row justify-content-center">
+              <div class="row justify-content-center hide-view-kept">
                 <div class="col-md-6">
                   <div class="row justify-content-center">
                     <div class="col-md-2 p-0 d-flex justify-content-end">
-                      <i class="mdi mdi-eye f-18 text-primary"></i>
+                      <i class="mdi mdi-eye f-22 text-primary"></i>
                     </div>
                     <div
                       class="
@@ -39,7 +40,7 @@
                     </div>
                     <div class="col-md-2 p-0 d-flex justify-content-end">
                       <i
-                        class="mdi mdi-alpha-k-box-outline f-18 text-primary"
+                        class="mdi mdi-alpha-k-box-outline f-22 text-primary"
                       ></i>
                     </div>
                     <div
@@ -69,12 +70,11 @@
                 <p>
                   {{ keep.description }}
                 </p>
-                <div class="border-bottom"></div>
               </div>
             </div>
             <div class="row">
               <span class="d-flex align-items-center justify-content-between">
-                <div class="mb-3">
+                <div class="d-flex align-items-center mt-3">
                   <select
                     class="selectable"
                     name="vaults"
@@ -88,14 +88,20 @@
                     </option>
                   </select>
                 </div>
-                <i
-                  v-if="keep.creatorId == account.id"
-                  class="button mdi mdi-delete f-24 selectable"
-                  @click="deleteKeep()"
-                ></i>
-
-                <div class="rounded p-2">
-                  <div>
+                <div class="d-flex align-items-center mt-3">
+                  <i
+                    v-if="keep.creatorId == account.id"
+                    title="Delete Keep"
+                    class="button mdi mdi-delete f-24 selectable"
+                    @click="deleteKeep()"
+                  ></i>
+                </div>
+                <div class="rounded p-2 d-flex align-items-center mt-3">
+                  <div
+                    :title="`View ${keep.creator?.name}'s Profile`"
+                    class="selectable"
+                    @click="goToProfile"
+                  >
                     <img
                       :src="account.picture"
                       alt="account photo"
@@ -152,13 +158,19 @@ export default {
       },
       async deleteKeep() {
         try {
-          Modal.getOrCreateInstance(document.getElementById('keep-modal')).hide()
-          await keepsService.deleteKeep(this.keep.id)
-          Pop.toast('Deleted Keep', 'error')
+          if (await Pop.confirm()) {
+            Modal.getOrCreateInstance(document.getElementById('keep-modal')).hide()
+            await keepsService.deleteKeep(this.keep.id)
+            Pop.toast('Deleted Keep', 'error')
+          }
         } catch (error) {
           logger.error(error)
           Pop.toast(error.message, 'error')
         }
+      },
+      goToProfile() {
+        Modal.getOrCreateInstance(document.getElementById('keep-modal')).hide()
+        router.push({ name: "Profile", params: { id: this.keep.creatorId } })
       },
       keep: computed(() => AppState.activeKeep),
       account: computed(() => AppState.account),
@@ -176,7 +188,7 @@ export default {
   height: 60vh;
 }
 .description {
-  height: 40vh;
+  height: 35vh;
 }
 select {
   margin-bottom: 10px;
@@ -194,6 +206,12 @@ select {
     display: none;
   }
 }
+@media all and (max-width: 1000px) {
+  .hide-view-kept {
+    display: none;
+  }
+}
+
 p {
   font-family: Verdana, Geneva, Tahoma, sans-serif;
   font-size: 20px;
